requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.12.0
python-dotenv==1.0.0
Pillow==10.2.0
pytesseract==0.3.8
PyMuPDF==1.23.26
python-docx==0.8.11
Werkzeug==3.0.0
pdfplumber==0.10.3
fastapi==0.111.0
uvicorn==0.29.0 